% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsFunctions.R
\name{arbammitJagsRealData}
\alias{arbammitJagsRealData}
\title{Jags code AR BAMMIT model}
\usage{
arbammitJagsRealData(
  data,
  Q = 1,
  mmu = 10,
  smu = 2,
  stheta = 1,
  a = 0.1,
  b = 0.1,
  nthin = 1,
  nburnin = 2000
)
}
\arguments{
\item{data}{A data frame containing the columns Yield, Genotype, Environment, Year and Block}

\item{Q}{number of components of interaction term}

\item{mmu}{Mean of the grand mean}

\item{smu}{sd of grand mean}

\item{a}{shape parameter of a Gamma distribution}

\item{b}{scale parameter of a Gamma distribution}

\item{nthin}{thinning rate}

\item{nburnin}{length of burn in}
}
\value{
The output of Jags function
}
\description{
A Jags code to estimated the parameters of AR BAMMIT model
}
