% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsFunctions.R
\name{bammitJags}
\alias{bammitJags}
\title{Jags code BAMMIT model}
\usage{
bammitJags(data, Q, mmu, smu, a, b, stheta = 1, nthin, nburnin)
}
\arguments{
\item{data}{A list containing the variables y and Bv.}

\item{Q}{number of components of interaction term}

\item{mmu}{Mean of the grand mean}

\item{smu}{sd of grand mean}

\item{a}{shape parameter of a Gamma distribution}

\item{b}{scale parameter of a Gamma distribution}

\item{nthin}{thinning rate}

\item{nburnin}{length of burn in}
}
\value{
The output of Jags function
}
\description{
A Jags code to estimated the parameters of BAMMIT model
}
