% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsFunctions.R
\name{josseJags}
\alias{josseJags}
\title{Jags code Jossi's model}
\usage{
josseJags(data, mmu, smu, sg, se, slambda, a, b, nthin = 1, nburnin = 2000)
}
\arguments{
\item{data}{A list containing the simulated values of y, the values of I, J, and Q and a mtrix od g and e.}

\item{mmu}{Mean of the grand mean}

\item{smu}{sd of grand mean}

\item{sg}{sd of genotypes}

\item{se}{sd of environments}

\item{slambda}{sd of $\lambda$}

\item{a}{shape parameter of a Gamma distribuion}

\item{b}{scale parameter of a Gamma distribuion}

\item{nthin}{thinning rate}

\item{nburnin}{length of burn in}
}
\value{
The output of Jags function
}
\description{
A Jags code to estimated the parameters of AMMI model followuing Josse's approach
}
