% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliarFunctions.R
\name{predictionBAMMIT}
\alias{predictionBAMMIT}
\title{Predicted values of BAMMIT model}
\usage{
predictionBAMMIT(model, data)
}
\arguments{
\item{model}{A object from Jags model}

\item{data}{A object from simulated data.}
}
\value{
A vector of predicted values of the response variable.
}
\description{
A function to calculate the predicted values of a BAMMIT model from a JAGS
}
