% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliarFunctions.R
\name{R2}
\alias{R2}
\title{R2}
\usage{
R2(pred, obs)
}
\arguments{
\item{true}{A vector of the true values of the response variable.}

\item{predicted}{A vector of the predicted values of the response variable.}
}
\value{
A number.
}
\description{
A function to calculate the R2.
}
