% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliarFunctions.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Square Error}
\usage{
RMSE(true, predicted)
}
\arguments{
\item{true}{A vector of the true values of the response variable.}

\item{predicted}{A vector of the predicted values of the response variable.}
}
\value{
A number.
}
\description{
A function to calculate the root mean square error.
}
