% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simBAMMIT}
\alias{simBAMMIT}
\title{Simulate data BAMMIT}
\usage{
simBAMMIT(
  V = 3,
  Q = 1,
  Bv = c(3, 2, 2),
  mu = 100,
  lambda = c(10),
  sb = 1,
  sB = 10,
  sy = 1
)
}
\description{
Simulate data BAMMIT
}
