% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateDataAmmi}
\alias{simulateDataAmmi}
\title{Simulate data from AMMI model}
\usage{
simulateDataAmmi(I, J, mu, sg, se, sy, lambda)
}
\arguments{
\item{I}{Number of genotypes}

\item{J}{Number of environments}

\item{mu}{Grand mean}

\item{sg}{standard deviation of genotypes}

\item{se}{standard deviation of environments}

\item{sy}{standard deviation of the model}

\item{lambda}{vector of values of $\lambda$}
}
\value{
A list containing the response y, genotypes, environments, bilinear term, Q, and a matrix of g and e.
}
\description{
A funtion to generate values of a AMMI model
}
