% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateDataBammit}
\alias{simulateDataBammit}
\title{Simulate data from Bayesian AMMIT model}
\usage{
simulateDataBammit(I, J, K, mu, sg, se, st, sy, lambda)
}
\arguments{
\item{I}{Number of genotypes.}

\item{J}{Number of environments.}

\item{K}{Number of terms of time.}

\item{mu}{Grand mean.}

\item{sg}{standard deviation of genotypes.}

\item{se}{standard deviation of environments.}

\item{st}{standard deviation of time.}

\item{sy}{standard deviation of the model.}

\item{lambda}{vector of values of $\lambda$.}
}
\value{
A list containing the response y, genotypes, environments, time,
bilinear term, Q, and a matrix of g, e and t.
}
\description{
A funtion to generate values of the BAMMIT model
}
